add middleware using jwt

+++ controller/product/deleteproduct.js
@@ -1,19 +1,27 @@
 const { Products } = require("../../models");
-
+const jwt = require("jsonwebtoken");
 async function deleteProduct(req, res) {
   try {
+    let header = req.headers.authorization.split("Bearer ")[1];
+    let userData = jwt.verify(header, "s3cr3t");
     const inputId = req.params.id;
-
     const checkProduct = await Products.findByPk(inputId);
 
     if (checkProduct) {
-      await Products.destroy({ where: { id: inputId } });
-      res.send(`Product dengan id ${inputId} berhasil di hapus`);
+      if (checkProduct.user_id == userData.id) {
+        await Products.destroy({ where: { id: inputId } });
+        res.send(`Product dengan id ${inputId} berhasil di hapus`);
+      } else {
+        res.status(403).json({
+          message: "Anda tidak bisa menghapus product yang bukan milik anda",
+        });
+      }
     } else {
       res.status(404).send("Produk tidak ditemukan");
     }
   } catch (error) {
     res.send(error);
+    return;
   }
 }
